Add mmap helper functions
Adds `align_down` which rounds a number down to multiple. Equivalent to
the `ALIGN_DOWN` macro in QEMU.

Adds `align_up` which rounds a number up to multiple. Equivalent to the
`ALIGN_UP` macro in QEMU`.

Adds `mmap_reserve` which reserves a new memory region of the requested
size to be used for mapping from the given fd (if any). This is
equivalent to `mmap_reserve` in QEMU.

Adds `mmap_activate` which activates memory in a reserved region from
the given fd (if any), to make it accessible.  This is equivalent to
`mmap_activate` in QEMU.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>